requirements.txt
# The-Magician Dependencies

# Authentication
bcrypt>=4.0.0
PyJWT>=2.8.0

# Configuration
PyYAML>=6.0

# Email (dev/production)
# For production, uncomment preferred service:
# sendgrid>=6.10.0
# mailgun>=0.1.1

# Terminal formatting
colorama>=0.4.6

# Data validation
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
python-dotenv>=1.0.0
